update some cast

+++ src/rendering/color.c
@@ -17,9 +17,9 @@ float	calculate_buffer_zone_x(t_data *data)
 	float	buff;
 
 	if (data->dir_x < 0)
-		buff = -0.2;
+		buff =(float)-0.2;
 	else
-		buff = 0.2;
+		buff = (float)0.2;
 	return (buff);
 }
 
@@ -28,9 +28,9 @@ float	calculate_buffer_zone_y(t_data *data)
 	float	buff;
 
 	if (data->dir_y < 0)
-		buff = -0.2;
+		buff = (float)-0.2;
 	else
-		buff = 0.2;
+		buff = (float)0.2;
 	return (buff);
 }
